% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecospat.boyce.R
\name{ecospat.boyce}
\alias{ecospat.boyce}
\title{ecospat.boyce}
\usage{
ecospat.boyce(
  fit,
  obs,
  nclass = 0,
  window.w = "default",
  res = 100,
  PEplot = TRUE,
  rm.duplicate = TRUE,
  method = "spearman"
)
}
\arguments{
\item{fit}{A vector or Raster-Layer containing the predicted suitability values}

\item{obs}{A vector containing the predicted suitability values or xy-coordinates (if fit is a Raster-Layer) of the validation points (presence records)}

\item{nclass}{Number of classes or vector with classes threshold. If nclass=0, Boyce index is calculated with a moving window (see next parameters)}

\item{window.w}{width of the moving window (by default 1/10 of the suitability range)}

\item{res}{resolution of the moving window (by default 100 focals)}

\item{PEplot}{if True, plot the predicted to expected ratio along the suitability class}

\item{rm.duplicate}{if TRUE, the correlation exclude successive duplicated values}

\item{method}{correlation method used to compute the boyce index}
}
\description{
Updated ecospat.boyce (ecospat package) function for calculating Boyce index as in Hirzel et al. 2006
}
\author{
Blaise Petitpierre (blaise.petitpierre@unil.ch)
}
