% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.absences.R
\name{nsdm.absences}
\alias{nsdm.absences}
\title{\code{nsdm.absences}}
\usage{
nsdm.absences(
  n = 10000,
  pres,
  taxon = character(),
  type = "po",
  rst_ref,
  rst_background_weight = NULL,
  set_max_npres_to_nabs = TRUE
)
}
\arguments{
\item{n}{Integer. Number of background points to be generated.}

\item{pres}{\code{sf} object. Spatial points dataset containing species occurrence data.}

\item{taxon}{Character. Target species name.}

\item{type}{Character. Specifies the type of species data:
\itemize{
\item \code{"po"}: Presence-only data
\item \code{"pa"}: Presence-absence data
}}

\item{rst_ref}{Raster (\code{SpatRaster}). Reference raster used in subsequent analyses.}

\item{rst_background_weight}{Raster (\code{SpatRaster}) or \code{NULL}. Background weight raster used for weighted pseudoabsence sampling. If \code{NULL}, sampling is random.}

\item{set_max_npres_to_nabs}{Logical. If \code{TRUE}, limits the maximum number of presence points to the number of background points.}
}
\value{
An \code{nsdm.pseudoabsences} object.
}
\description{
Prepare background (pseudoabsence) points for species distribution modeling.
}
\author{
Antoine Adde (\email{antoine.adde@eawag.ch})
}
