% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.covinfo.R
\name{nsdm.covinfo}
\alias{nsdm.covinfo}
\title{nsdm.covinfo}
\usage{
nsdm.covinfo(cov_path, save_path)
}
\arguments{
\item{cov_path}{A character string specifying the directory where covariates are stored.}

\item{save_path}{A character string specifying the destination path to save the covariate information table.}
}
\value{
A \code{.psv} file containing the covariate information table.
}
\description{
Generate Covariate Information Table
}
\details{
This function generates a covariate information table based on the provided covariate paths and names.
}
\author{
Antoine Adde (antoine.adde@eawag.ch)
}
