% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.ensemble.R
\name{nsdm.ensemble}
\alias{nsdm.ensemble}
\title{nsdm.ensemble}
\usage{
nsdm.ensemble(
  model_names,
  species_name,
  level = NA,
  nesting_name = NA,
  scenar_name = NA,
  period_name = NA,
  map_path,
  score_path = NULL,
  discthre = NULL,
  weighting = FALSE,
  weight_metric = "Score"
)
}
\arguments{
\item{model_names}{A character vector indicating the modelling algorithms to ensemble.}

\item{species_name}{A character string indicating the name of the taxon.}

\item{level}{A character string indicating the name of the level considered (e.g., "glo" or "reg").}

\item{nesting_name}{A character string indicating the name of the nesting strategy for which predicted values are provided.}

\item{scenar_name}{A character string indicating the name of the scenario for which predicted values are provided.}

\item{period_name}{A character string indicating the name of the period for which predicted values are provided.}

\item{map_path}{A character string indicating the path where algorithm-specific prediction surfaces are stored.}

\item{score_path}{A character string indicating the path where algorithm-specific evaluation metrics are stored.}

\item{discthre}{A numeric value specifying the threshold for \code{weight_metric}, below which an algorithm is discarded.
If \code{NULL}, no models are discarded.}

\item{weighting}{Logical. If \code{TRUE}, individual algorithms are weighted by \code{weight_metric}.}

\item{weight_metric}{A character string indicating the name of the metric used for weighting or discarding models.}
}
\value{
A named list of two \code{SpatRaster} objects:
\describe{
\item{"ensemble"}{A \code{SpatRaster} representing the average ensemble prediction.}
\item{"ensemble_cv"}{A \code{SpatRaster} representing the coefficient of variation of the ensemble.}
}
}
\description{
Ensemble prediction surfaces obtained by individual algorithms.
}
\author{
Antoine Adde (antoine.adde@eawag.ch)
}
