% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.eval.R
\name{nsdm.eval}
\alias{nsdm.eval}
\title{nsdm.eval}
\usage{
nsdm.eval(x, models, sets, level, ncores = 1L, tmp_path)
}
\arguments{
\item{x}{An \code{nsdm.pseudoabsences} object that holds the master covariates, SIDs, and presence values}

\item{models}{An \code{nsdm.fit} object produced by your fitting step, models are in \code{models@fits}}

\item{sets}{A list of replicates, each replicate contains \code{reg_train}, \code{reg_test}, \code{glo_train}, \code{glo_test}}

\item{level}{Character string, either \code{"glo"} or \code{"reg"}, selects which SIDs to use for testing}

\item{ncores}{Integer, number of cores used by \code{parallel::mclapply}, default is \code{1}}

\item{tmp_path}{Character path for LightGBM model files, default is \code{tempdir()}}
}
\value{
An object of class \code{nsdm.evaluation} with performance per model tag and replicate
}
\description{
Parallel evaluation of fitted models with common SDM metrics.
}
\author{
Antoine Adde (antoine.adde@eawag.ch)
}
