% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.evaluation.R, R/nsdm.evaluation_class.R
\docType{class}
\name{nsdm.evaluation-class}
\alias{nsdm.evaluation-class}
\alias{nsdm.evaluation}
\title{nsdm.evaluation}
\arguments{
\item{x}{A nsdm.fit object}

\item{level}{A character string indicating the evaluated level (REG or GLO)}

\item{ncores}{Number of cores to be used during parallel operations}

\item{tmp_path}{A character string indicating the path where to store gbm temporary outputs}
}
\value{
An object of class 'nsdm.evaluation'
}
\description{
Parallel evaluation of nsdm.fit models with a suite of model assessment metrics

Not to be called directly by the user
}
\author{
Philipp Brun (philipp.brun@wsl.ch)
}
