% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.fitting3.R
\name{nsdm.flex3}
\alias{nsdm.flex3}
\title{nsdm.flex3}
\usage{
nsdm.flex3(
  x = numeric(),
  replicatetype = character(),
  evaluationdomain = character(),
  reps,
  mod_args = list(),
  path = NA,
  level,
  ncores,
  tmp_path
)
}
\arguments{
\item{x}{A nsdm.pseudoabsences object}

\item{replicatetype}{A charachter string indicating how should replicates be generated? may be 'none', 'splitsample', or 'clustered_splitsample'}

\item{evaluationdomain}{A charachter string indicating the spatial domain for evaluation? may be 'regionalonly', or 'all'}

\item{reps}{Numeric number of replicates}

\item{mod_args}{List of class 'multi.input' containing information on models to be fitted}

\item{path}{where to save? (not implemented yet)}

\item{level}{A charachter string indicating the level evaluated (e.g. glo or reg)}

\item{ncores}{Number of cores to be used during parallel operations}

\item{tmp_path}{Character indicating the path where to store temporary outputs}
}
\value{
A nsdm.fit object including slots for meta info, testing data for evaluation, and model objects
}
\description{
Model fitting (parallel version)
}
\author{
Philipp Brun (philipp.brun@wsl.ch) and Antoine Adde (antoine.adde@eawag.ch)
}
