% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.loadthis.R
\name{nsdm.loadthis}
\alias{nsdm.loadthis}
\title{nsdm.loadthis}
\usage{
nsdm.loadthis(model_name = NULL, species_name = NULL, tag = NULL, read_path)
}
\arguments{
\item{model_name}{A character string indicating the name of the modelling alogirthm related to the target object to be loaded}

\item{species_name}{A character string  indicating the name of the species related to the target object to be loaded}

\item{tag}{A character string indicating the tag for the hyperparameter X modelling algorithm combination of the target object to be loaded}

\item{read_path}{A character string indicating the upper path where the target object is stored (above model and species)}
}
\value{
The desired nsdm.ine object
}
\description{
Easy loading of nsdm.ine objects
}
\author{
Antoine Adde (aadde@unil.ch)
}
