% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.map.R
\name{nsdm.map}
\alias{nsdm.map}
\title{nsdm.map}
\usage{
nsdm.map(
  template,
  nona_ix,
  pred,
  species_name,
  model_name,
  level = NA,
  nesting_name = NA,
  scenar_name = NA,
  period_name = NA
)
}
\arguments{
\item{template}{An empty raster layer to fill with predicted values.}

\item{nona_ix}{Numeric indices of cells that should be filled with predicted values.}

\item{pred}{A list where each element corresponds to a modeling algorithm and contains vectors of fitted values.}

\item{species_name}{A character string indicating the name of the taxon for which models are fitted.}

\item{model_name}{A character string indicating the name of the modeling algorithm used.}

\item{level}{A character string indicating the spatial level considered (e.g., "glo" for global, "loc" for local).}

\item{nesting_name}{A character string indicating the nesting strategy for which predicted values are provided.}

\item{scenar_name}{A character string indicating the scenario for which predicted values are provided.}

\item{period_name}{A character string indicating the time period for which predicted values are provided.}
}
\value{
A list of rasters filled with predicted values (for ESM models) or a single raster (for other models).
}
\description{
Fill an empty raster template with predicted values.
}
\author{
Antoine Adde (antoine.adde@eawag.ch)
}
