% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.multi.R
\name{nsdm.multi}
\alias{nsdm.multi}
\title{nsdm.multi}
\usage{
nsdm.multi(mod, args, tag = "", step = FALSE, weight = FALSE)
}
\arguments{
\item{mod}{A character with the name of the function to be called. E.g. "gam"}

\item{args}{A list with arguments to be passed to the function specified in mod}

\item{tag}{A character with name for model set-up}
}
\value{
A multi.input object that efficiently stores model specifications
}
\description{
Define settings for function that should be supplied to nsdm.flex
}
\author{
Philipp Brun (philipp.brun@wsl.ch)
}
