% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.pseuabs.R
\name{nsdm.pseuabs}
\alias{nsdm.pseuabs}
\title{nsdm.pseuabs}
\usage{
nsdm.pseuabs(
  n = 10000,
  pres,
  taxon = character(),
  type = "po",
  rst_ref,
  set_max_npres_to_nabs = TRUE
)
}
\arguments{
\item{n}{Number of background points to be generated}

\item{pres}{Spatial point object containing species occurrence data}

\item{taxon}{A character string for the target species}

\item{type}{A character string indicating the type of species data ("po"=presence only; "pa"=presence absence)}

\item{rst_ref}{Reference raster used in subsequent analyses}

\item{set_max_npres_to_nabs}{Logical (TRUE FALSE) for setting the maximal number of presences to the number of background points}
}
\value{
A nsdm.pseudoabsences object
}
\description{
Prepare background (pseudoabsences) points
}
\author{
Antoine Adde (aadde@unil.ch) and Philipp Brun (philipp.brun@wsl.ch)
}
