% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.pseudoabsences_class.R
\docType{class}
\name{nsdm.pseudoabsences-class}
\alias{nsdm.pseudoabsences-class}
\alias{nsdm.pseudoabsences}
\title{Generate an S4 class to store pseudoabsences data}
\description{
Not to be called directly by the user
}
\author{
Philipp Brun (philipp.brun@wsl.ch)
}
