% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.retrieve4pred.R
\name{nsdm.retrieve4pred}
\alias{nsdm.retrieve4pred}
\title{nsdm.retrieve4pred}
\usage{
nsdm.retrieve4pred(
  covstk,
  scaleparam,
  mask = NULL,
  observational = NULL,
  obsval = NULL
)
}
\arguments{
\item{covstk}{A RasterStack containing the covariate layers used for prediction.}

\item{scaleparam}{A data frame or matrix containing the scaling parameters to be reapplied.}

\item{mask}{A character string specifying the path to an \code{.rds} file containing a raster where cells with a value of \code{0} will be masked during prediction.}

\item{observational}{A character vector specifying the names of "observational" covariates, which will be assigned a constant value (\code{obsval}) for prediction.}

\item{obsval}{A character string indicating the strategy for setting the common value for observational covariates. Options: \code{"zero"}, \code{"median"}, or \code{"max"}.}
}
\value{
A list containing:
\item{data}{A \code{data.frame} containing the covariate values for prediction.}
\item{valid_cells}{An index vector identifying non-NA cells in the original raster stack.}
}
\description{
Prepare covariate data for prediction.
}
\author{
Antoine Adde (\email{antoine.adde@eawag.ch})
}
