% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.savemap.R
\name{nsdm.savemap}
\alias{nsdm.savemap}
\title{nsdm.savemap}
\usage{
nsdm.savemap(maps, save_path, species_name, model_name = NULL, format = "both")
}
\arguments{
\item{maps}{A nsdm.map output (list of rasters filled with predicted values)}

\item{save_path}{A character string indicating the path where output maps will be saved}

\item{species_name}{A character string indicating the name of the target taxon}

\item{model_name}{A character string indicating the name of the target modelling alogirthm}

\item{format}{A character string indicating the saving format(s) ("rds", "tif" or "both")}
}
\value{
Saved maps in both .rds and/or .tif formats
}
\description{
Save individual predictive maps
}
\author{
Antoine Adde (aadde@unil.ch)
}
