% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.summary.R
\name{nsdm.summary}
\alias{nsdm.summary}
\title{nsdm.summary}
\usage{
nsdm.summary(object)
}
\arguments{
\item{object}{A nsdm.evaluation object to be summarized}
}
\description{
Summarize nsdm.evaluation objects
}
\author{
Philipp Brun (philipp.brun@wsl.ch) and Antoine Adde (antoine.adde@eawag.ch)
}
