% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.varimp.R
\name{nsdm.varimp}
\alias{nsdm.varimp}
\title{nsdm.varimp}
\usage{
nsdm.varimp(models)
}
\arguments{
\item{models}{A nsdm.fit object containing fitted model(s)}
}
\value{
A list containing the values of the relative importance for each individual covariate and model
}
\description{
Covariate importance computation
}
\author{
Antoine Adde (antoine.adde@eawag.ch)
}
