% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.preva_meta.R
\name{preva.meta}
\alias{preva.meta}
\title{Prediction-evaluation meta.info function: generate meta information for prediction and evaluation}
\usage{
preva.meta(env = parent.frame(), type = character())
}
\description{
Not to be called directly by the user
}
\author{
Philipp Brun (philipp.brun@wsl.ch)
}
