% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.all_metrics_function.R
\name{sens}
\alias{sens}
\title{Evaluation function (define some confusion matrix indices)}
\usage{
sens(a, c)
}
\description{
Evaluation function (define some confusion matrix indices)
}
\author{
Philipp Brun (philipp.brun@wsl.ch) and Antoine Adde (aadde@unil.ch)
}
