% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.covinfo.R
\name{nsdm.covinfo}
\alias{nsdm.covinfo}
\title{nsdm.covinfo}
\usage{
nsdm.covinfo(cov_path, save_path, time_cov, focal_cov)
}
\arguments{
\item{cov_path}{A character string specifying the root directory where covariates are stored.}

\item{save_path}{A character string specifying the destination path to save the \code{.xlsx} covariate information table.}

\item{time_cov}{A character vector specifying the categories or datasets that include temporally dynamic covariates.}

\item{focal_cat}{A character vector specifying the categories or datasets that include covariates with focal statistics.}
}
\value{
An \code{.xlsx} file containing the covariate information table.
}
\description{
Generate Covariate Information Table
}
\details{
This function generates a covariate information table in \code{.xlsx} format based on the provided covariate paths and names.

This function processes the covariate paths to identify dynamic and focal covariates, and saves the resulting information table in Excel format.
}
\author{
Antoine Adde
}
