% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.evaluation_class.R
\docType{class}
\name{nsdm.evaluation-class}
\alias{nsdm.evaluation-class}
\alias{nsdm.evaluation}
\title{Generate an S4 class to store evaluation data}
\description{
Not to be called directly by the user
}
\author{
Philipp Brun (philipp.brun@wsl.ch)
}
