% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsdm.fitting3.R
\name{nsdm.flex3}
\alias{nsdm.flex3}
\title{nsdm.flex3}
\usage{
nsdm.flex3(
  x = numeric(),
  pa = numeric(),
  env_vars = data.frame(),
  taxon = character(),
  replicatetype = character(),
  reps,
  mod_args = list(),
  timer = FALSE,
  save = FALSE,
  project = NA,
  path = NA,
  level,
  ncores,
  tmp_path
)
}
\arguments{
\item{x}{A nsdm.pseudoabsences object}

\item{pa}{A vector containing 0/1 presence/absence values}

\item{env_vars}{A data.frame containing environmental predictors}

\item{taxon}{A character string of the taxon for which models are fitted}

\item{replicatetype}{A charachter string indicating how should replicates be generated? may be 'none', 'splitsample',
'cv' 'block-cv'}

\item{reps}{Numeric number of replicates}

\item{mod_args}{List of class 'multi.input' containing information on models to be fitted}

\item{timer}{Logical (TRUE or FALSE) indicating if model computation time is to be saved}

\item{save}{should the model be saved in a structured way? (not implemented yet)}

\item{project}{character indicating the name of the project within which the models are run (later used to define saving directories)}

\item{path}{where to save? (not implemented yet)}

\item{level}{A charachter string indicating the level evaluated (e.g. glo or reg)}

\item{ncores}{Number of cores to be used during parallel operations}

\item{tmp_path}{Character indicating the path where to store temporary outputs}

\item{strata}{A numeric vector of the same length as observations with integers separating cross validation replicates (used when replicatetype='block-cv')}
}
\value{
A nsdm.fit object including slots for meta info, testing data for evaluation, and model objects
}
\description{
Model fitting (parallel version)
}
\author{
Philipp Brun (philipp.brun@wsl.ch) and Antoine Adde (antoine.adde@eawag.ch)
}
